Fix funnel EOS handling and wrong unittest
When no data is coming from sinkpads and eos events
arrived at one of the sinkpad, funnel forwards the EOS
event to downstream. It forwards the EOS because lastsink pad
is NULL. Also the unit testcase of the funnel is not checking
the correct behavior as it should. The unit test case should
fail if one of the sink pad has already EOS present on it and
we are trying to push one more EOS.

https://bugzilla.gnome.org/show_bug.cgi?id=731716

Change-Id: I6b8e38a65b1627bbf27a3d4448eb5e0c66e49034
Reviewed-on: https://gpro.lgsvl.com/67147
Tested-by: HoonHee Lee <hoonhee.lee@lge.com>
Reviewed-by: Jeongseok Kim <jeongseok.kim@lge.com>
Tested-by: Jeongseok Kim <jeongseok.kim@lge.com>