funnel: Fix for racy EOS event handling
When eos events are forwarded simultaneouly from two sinkpads on
funnel, it doesnot forward the eos to sourcepad. The reason is
sticky events are stored after the event callbacks are returned.
Therefore while one is about to store the sticky events on the its
sinkpad, other sinkpad starts checking for the eos events on all other
sinkpads and assumes eos is not present yet.

https://bugzilla.gnome.org/show_bug.cgi?id=732851

Change-Id: I7e4fa856735ce02aec31e01cfc0ed705f851c0c2
Reviewed-on: https://gpro.lgsvl.com/67148
Tested-by: HoonHee Lee <hoonhee.lee@lge.com>
Reviewed-by: Jeongseok Kim <jeongseok.kim@lge.com>
Tested-by: Jeongseok Kim <jeongseok.kim@lge.com>